xfree86: ddc: move struct cea_data_block into private header
Not used by any external drivers, so no need to keep it in public header.
