+++ hw/xfree86/modes/xf86Crtc.c
@@ -33,6 +33,7 @@
 #include "dix/screen_hooks_priv.h"
 #include "randr/randrstr_priv.h"
 
+#include "edid_priv.h"
 #include "xf86_priv.h"
 #include "xf86DDC_priv.h"
 #include "xf86Config.h"
